+++ borax/__init__.py
@@ -1,4 +1,4 @@
 # coding=utf8
 
-__version__ = '1.3.1'
+__version__ = '1.4.0'
 __author__ = 'kinegratii'